libft

+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: hstolten <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/20 06:53:09 by hstolten          #+#    #+#             */
-/*   Updated: 2018/11/20 06:53:16 by hstolten         ###   ########.fr       */
+/*   Updated: 2018/12/11 17:25:55 by hstolten         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int     ft_strncmp(const char *s1, const char *s2, size_t n)
+int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] != '\0' && s2[i] != '\0' && i < n)
